requirements.txt
numpy>=1.18.4
opencv-python~=4.2.0.34
future~=0.18.2
pillow~=7.1.2
pytorch~=1.1.0
torchvision~=0.6.0
cudatoolkit~=9.0
scipy>=1.4.1
scikit-image>=0.15.0
matplotlib>=3.1.0
imageio>=2.8.0
pydicom>=1.3.0
pandas>=0.25.0
scikit-learn~=0.21.2
h5py>=2.8.0
pyprind>=2.11.2
PyYAML>=5.1.1
tensorboard
psutil